Updated with suggested changes Fixed some minor bugs

+++ game/building/src/widgets/BuildPanel/widgets/Blueprints/services/session/requester.ts
@@ -5,21 +5,27 @@
  */
 import { client, restAPI, buildUIMode } from 'camelot-unchained';
 import {Blueprint, UNCLASSIFIED} from '../../lib/Blueprint';
+import faker from './requester_fake';
 
 class BlueprintLoader {
-
+  private win: any = window;
+  private fake: boolean = (this.win.cuAPI == null);
 
   public loadBlueprints(callback: (print: Blueprint) => void) {
+    if (this.fake) {
+      return faker.loadBlueprints(callback);
+    }
+
     client.OnNewBlueprint((index: number, fullName: string) => {
       //special case, this is not a real blueprint
       if (fullName == "Small Control Island")
         return;
 
-      let catAndName = this.splitCategoryAndName(fullName);
-      let blueprint = {
+      const catAndName = this.splitCategoryAndName(fullName);
+      const blueprint = {
         index: index,
         id: fullName,
-        name: catAndName.name,
+        name: fullName,
         category: catAndName.cat,
       } as Blueprint;
       callback(blueprint);
@@ -28,6 +34,10 @@ class BlueprintLoader {
   }
 
   public loadBlueprintIcon(blueprint: Blueprint, callback: (blueprint: Blueprint, icon: string) => void) {
+    if (this.fake) {
+      return faker.loadBlueprintIcon(blueprint, callback);
+    }
+
     restAPI.getBlueprintIcon(blueprint.index).then((icon: string): void => {
       callback(blueprint, icon);
     }, (): void => {
@@ -35,64 +45,79 @@ class BlueprintLoader {
     })
   }
 
-  public listenForModeChange(callback: (selected: boolean) => void) {
-    
-  /* mode is defined as a boolean, it is actually a number */
+  public listenForSelectionModeChange(callback: (selected: boolean) => void) {
+    if (this.fake) {
+      return faker.listenForSelectionModeChange(callback);
+    }
 
     client.OnBuildingModeChanged((mode: number) => {
       callback(mode==buildUIMode.BLOCKSELECTED);
-    });
+    });    
   }
 
   public listenForCopy(callback: () => void) {
+    if (this.fake) {
+      return faker.listenForCopy(callback);
+    }
+
     client.OnCopyBlueprint(() => {
       callback();
-    });
-    client.OnBlueprintSelected(() => {
-      console.log("OnBlueprintSelected");
     });
   }
 
   public changeBlueprintSelection(blueprint: Blueprint) {
+    if (this.fake) {
+      return faker.changeBlueprintSelection(blueprint);
+    }
+
     client.SelectBlueprint(blueprint.index);
   }
 
-  public select(blueprint: Blueprint)
-  {
+  public select(blueprint: Blueprint) {
+    if (this.fake) {
+      return faker.select(blueprint);
+    }
+
     client.SelectBlueprint(blueprint.index);
   }
 
   public save(name: string) {
+    if (this.fake) {
+      return faker.save(name);
+    }
+
     client.SaveBlueprint(name);
   }
 
   public remove(blueprint: Blueprint) {
+    if (this.fake) {
+      return faker.remove(blueprint);
+    }
+
     //no feedback on this delete, we just call it and cross our fingers
     client.DeleteLocalBlueprint(blueprint.id);
   }
 
   public copy() {
+    if (this.fake) {
+      return faker.copy();
+    }
+
     client.CopyBlueprint();
   }
 
   public paste() {
+    if (this.fake) {
+      return faker.paste();
+    }
+    
     client.PasteBlueprint();
   }
 
-  private listenBlueprints = (): void => {
-    client.OnDownloadBlueprints((charId: string) => {
-      console.log("OnDownloadBlueprints " + charId);
-    });
-
-    client.OnUploadBlueprint((charId: string, name: string, data: any) => {
-      console.log("OnUploadBlueprint");
-    });
-  }
-
   private splitCategoryAndName(fullName: string) {
-    var category = UNCLASSIFIED;
-    var name = fullName;
-    var index = fullName.indexOf('-');
+    let category = UNCLASSIFIED;
+    let name = fullName;
+    const index = fullName.indexOf('-');
     if (index > 0) {
       category = fullName.substring(0, index);
       name = fullName.substring(index + 1);
